Fixed warnings
Many warnings are fixed. Most (all?) of the following types:
a) Not all code paths return a value (after an assert)
b) Implicit casts of integer value
c) Unused variable

None of these did reveal bugs. Nevertheless, good to have the code cleaned up.

Wrt to the usage of integers did more than fix warnings. More generally changed declarations to better reflecting the semantics by switching to unsigned integers where appropriate.

+++ control/DirectoryViewToolbarControl.h
@@ -10,10 +10,10 @@
   IBOutlet NSSegmentedControl  *zoomControls;
   IBOutlet NSSegmentedControl  *focusControls;
   
-  int  zoomInSegment;
-  int  zoomOutSegment;
-  int  focusUpSegment;
-  int  focusDownSegment;
+  NSUInteger  zoomInSegment;
+  NSUInteger  zoomOutSegment;
+  NSUInteger  focusUpSegment;
+  NSUInteger  focusDownSegment;
 
   DirectoryViewControl  *dirViewControl;
