+++ test/strings/StringPrefixTest.m
@@ -11,8 +11,8 @@
 
 
 - (BOOL) testString:(NSString *)string matches:(NSString *)match {
-  int  stringLen = [string length];
-  int  matchLen = [match length];
+  NSUInteger  stringLen = [string length];
+  NSUInteger  matchLen = [match length];
   
   if (stringLen < matchLen) {
     return NO;
